feat: adicionando Operador e alterando algumas regras de negócio

+++ src/main/java/br/com/fiap/hackathon/reservafacil/service/impl/MedicamentoServiceImpl.java
@@ -1,14 +1,20 @@
 package br.com.fiap.hackathon.reservafacil.service.impl;
 
 import br.com.fiap.hackathon.reservafacil.exception.medicamento.MedicamentoNaoEncontradoException;
+import br.com.fiap.hackathon.reservafacil.exception.usuario.AcessoNegadoException;
 import br.com.fiap.hackathon.reservafacil.mapper.MedicamentoMapper;
 import br.com.fiap.hackathon.reservafacil.model.Medicamento;
+import br.com.fiap.hackathon.reservafacil.model.Operador;
 import br.com.fiap.hackathon.reservafacil.model.Prestador;
+import br.com.fiap.hackathon.reservafacil.model.Usuario;
 import br.com.fiap.hackathon.reservafacil.model.dto.medicamento.AtualizarMedicamentoRequestDTO;
 import br.com.fiap.hackathon.reservafacil.model.dto.medicamento.CadastrarMedicamentoRequestDTO;
 import br.com.fiap.hackathon.reservafacil.model.dto.medicamento.MedicamentoResponseDTO;
 import br.com.fiap.hackathon.reservafacil.repository.MedicamentoRepository;
+import br.com.fiap.hackathon.reservafacil.repository.OperadorRepository;
+import br.com.fiap.hackathon.reservafacil.security.SecurityService;
 import br.com.fiap.hackathon.reservafacil.service.MedicamentoService;
+import br.com.fiap.hackathon.reservafacil.service.OperadorService;
 import br.com.fiap.hackathon.reservafacil.service.PrestadorService;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,17 +25,30 @@ import java.util.UUID;
 
 @Service
 public class MedicamentoServiceImpl implements MedicamentoService {
-
     @Autowired
     private MedicamentoRepository medicamentoRepository;
 
     @Autowired
     private PrestadorService prestadorService;
 
+    @Autowired
+    private SecurityService securityService;
+
+    @Autowired
+    private OperadorService operadorService;
+
     @Override
     @Transactional
     public MedicamentoResponseDTO cadastrarMedicamento(CadastrarMedicamentoRequestDTO dto) {
         Prestador prestador = prestadorService.buscarPrestadorPorId(dto.prestadorId());
+
+        Usuario usuarioLogado = securityService.obterUsuarioLogado();
+        Operador operador = operadorService.buscarPorCns(usuarioLogado.getCns());
+
+        if(!operador.getPrestador().getId().equals(prestador.getId())){
+            throw new AcessoNegadoException("Você não pode cadastrar um medicamento em um prestador que você não está associado.");
+        }
+
         Medicamento medicamento = MedicamentoMapper.toMedicamento(dto);
         medicamento.setPrestador(prestador);
         return MedicamentoMapper.toMedicamentoResponseDTO(medicamentoRepository.save(medicamento));
@@ -39,6 +58,13 @@ public class MedicamentoServiceImpl implements MedicamentoService {
     @Transactional
     public MedicamentoResponseDTO atualizarMedicamento(UUID id, AtualizarMedicamentoRequestDTO dto) {
         Medicamento medicamento = buscarMedicamento(id);
+        Usuario usuarioLogado = securityService.obterUsuarioLogado();
+        Operador operador = operadorService.buscarPorCns(usuarioLogado.getCns());
+
+        if(!operador.getPrestador().getId().equals(medicamento.getPrestador().getId())){
+            throw new AcessoNegadoException("Você não pode atualizar um medicamento de um prestador que você não está associado.");
+        }
+
         if (dto.nome() != null && !dto.nome().isEmpty()) {
             medicamento.setNome(dto.nome());
         }
@@ -54,7 +80,14 @@ public class MedicamentoServiceImpl implements MedicamentoService {
     @Override
     @Transactional
     public void excluirMedicamento(UUID id) {
-        buscarMedicamento(id);
+        Medicamento medicamento = buscarMedicamento(id);
+        Usuario usuarioLogado = securityService.obterUsuarioLogado();
+        Operador operador = operadorService.buscarPorCns(usuarioLogado.getCns());
+
+        if(!operador.getPrestador().getId().equals(medicamento.getPrestador().getId())){
+            throw new AcessoNegadoException("Você não pode deletar um medicamento de um prestador que você não está associado.");
+        }
+
         medicamentoRepository.deleteById(id);
     }
